requirements.txt
django
beautifulsoup4
python-docx